pyproject.toml
[tool.poetry]
name = "mnemorphon"
version = "0.1.0"
description = "Exemplar language modeling in python"
authors = ["Your Name <you@example.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = "<2.0.0"
matplotlib = "^3.9.2"
librosa = "^0.10.2.post1"
tslearn = "^0.6.3"
tqdm = "^4.67.1"
scikit-learn = "^1.5.2"
pandas = "^2.2.3"
llvmlite = "^0.43.0"
torch = "2.2.2"
torchaudio = "2.2.2"
nemo-toolkit = {extras = ["tts"], version = "^2.0.0"}
jupyter = "^1.1.1"
notebook = "^7.2.2"
ipython = "^8.30.0"
zeyrek = "^0.1.3"
h5py = "^3.12.1"
pyyaml = "^6.0.2"
ipdb = "^0.13.13"

[[tool.poetry.source]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
